// src/components/TheyAreWithUs/TheyAreWithUs.types.ts
export enum TheyAreWithUsEnTags {
  All = 'All',
  Policy = 'Policy',
  Culture = 'Culture',
  Bussiness = 'Bussiness',
  Science = 'Science',
  Press = 'Press',
}

export enum TheyAreWithUsUkTags {
  All = 'All',
  Policy = 'Політика',
  Culture = 'Культура',
  Bussiness = 'Бізнес',
  Science = 'Наука',
  Press = 'Преса',
}

export enum TheyAreWithUsLocationEnTags {
  All = 'All',
  Usa = 'Usa',
  UnitedKingdom = 'United Kingdom',
  Germany = 'Germany',
  Ukraine = 'Ukraine',
  Portugal = 'Portugal',
  France = 'France',
  More = 'More',
}

export enum TheyAreWithUsLocationUkTags {
  All = 'Всі',
  Usa = 'Сша',
  UnitedKingdom = 'Велика Британія',
  Germany = 'Німеччина',
  Ukraine = 'Україна',
  Portugal = 'Португалія',
  France = 'Франція',
  More = 'Більше',
}

export type TheyAreWithUsPersonType = {
  photo: string;
  name: string;
  occupation: string;
  text?: string;
};

export type TheyAreWithUsItemType = {
  person: TheyAreWithUsPersonType;
  tags: TheyAreWithUsEnTags[] | TheyAreWithUsUkTags[];
  locationTags: TheyAreWithUsLocationEnTags[] | TheyAreWithUsLocationUkTags[];
};